requirements.txt
lightning==2.2.0
omegaconf==2.3.0
jaxtyping
typeguard
git+https://github.com/KAIR-BAIR/nerfacc.git@v0.5.2
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
diffusers==0.21.2
transformers==4.25.1
accelerate
opencv-python
tensorboard
matplotlib
imageio
imageio[ffmpeg]
trimesh
git+https://github.com/NVlabs/nvdiffrast.git
libigl
xatlas
trimesh[easy]
networkx
pysdf

# deepfloyd
xformers==0.0.21
bitsandbytes
sentencepiece
safetensors
huggingface_hub
libigl
xatlas
trimesh
networkx
PyMCubes
wandb

# for zero123
einops
kornia
taming-transformers-rom1504
git+https://github.com/openai/CLIP.git
open-clip-torch==2.24.0
rembg